+++ main.go
@@ -32,24 +32,15 @@ func main() {
 	}
 	cli.RootCommandHelpTemplate = `NAME:
 	{{.Name}} - {{.Usage}}
- USAGE:
-	{{.HelpName}} {{if .VisibleFlags}}[global options]{{end}}{{if .Commands}} command [command options]{{end}} {{if .ArgsUsage}}{{.ArgsUsage}}{{else}}[arguments...]{{end}}
-	{{if len .Authors}}
- AUTHOR:
-	{{range .Authors}}{{ . }}{{end}}
-	{{end}}{{if .Commands}}
- COMMANDS:
- {{range .Commands}}{{if not .HideHelp}}   {{join .Names ", "}}{{ "\t"}}{{.Usage}}{{ "\n" }}{{end}}{{end}}{{end}}{{if .VisibleFlags}}
- GLOBAL OPTIONS:
-	{{range .VisibleFlags}}{{.}}
-	{{end}}{{end}}{{if .Copyright }}
- COPYRIGHT:
-	{{.Copyright}}
-	{{end}}{{if .Version}}
- VERSION:
-	{{.Version}}
-	{{end}}
- `
+USAGE:
+	{{if .VisibleFlags}}[global options]{{end}}[target|completion [subcommand]]
+{{if .Commands}}
+COMMANDS:{{range .Commands}}{{if not .HideHelp}}
+	{{join .Names ", "}}{{ "\t"}}{{.Usage}}{{end}}{{end}}{{end}}
+{{if .VisibleFlags}}
+GLOBAL OPTIONS:{{range .VisibleFlags}}
+	{{.}}{{end}}{{end}}
+`
 
 	app := cli.Command{
 		Name:                   "webscan",