+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bug Fixes
 
 - Fixed bug where install.sh would fail to verify the checksum of the downloaded file
+- Fixed bug where help would not show usage samples
 - Fixed bug where `completion` subcommand would be shown twice in help
 
 ## v4.0.0